+++ client/src/components/page-template/PageTemplate.js
@@ -35,6 +35,7 @@ import {
   JsonReference,
   FileStorage,
   Settings,
+  VlanIbm,
 } from "@carbon/icons-react";
 import f5 from "../../images/f5.png";
 import { arraySplatIndex, contains, getObjectFromArray } from "lazy-z";
@@ -87,6 +88,7 @@ const navIcons = {
   IbmPowerVs: IbmPowerVs,
   IbmPowerVsInstance: VirtualMachine,
   IbmPowerVsVolumes: FileStorage,
+  VlanIbm: VlanIbm,
 };
 
 let pageOrder = [