+++ client/src/lib/forms/wizard.js
@@ -21,6 +21,7 @@ function wizard(wizardJson, json) {
       prefix: "wizard",
       tags: ["craig", "wizard"],
       fs_cloud: wizardJson.fs_cloud,
+      enable_classic: false,
       enable_power_vs: wizardJson.enable_power_vs,
       power_vs_zones: wizardJson.power_vs_zones || [],
       dynamic_subnets: !wizardJson.use_f5,